+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Update for 2024-06-14
 
 - force apply vae config on model load
+- restructure api examples: `cli/api-*`
+- fix control second pass resize
+- fix api face-hires
 
 ## Update for 2024-06-13